+++ src/bin.cpp
@@ -5,7 +5,7 @@
 #include <iostream>
 #include <functional> 
 
-Bin::Bin(const std::string& name, float w, float h, float d) 
+Bin::Bin(const std::string& name, long w, long h, long d) 
     : Box(name, w, h, d) {
 }
 
@@ -21,7 +21,7 @@ void Bin::addItem(Item& item) {
     items.push_back(std::ref(item));
 }
 
-float Bin::scoreRotation(const Item& item, int rotationType) const {
+float Bin::scoreRotation(const Item& item, long rotationType) const {
     Item rotatedItem = item;
     rotatedItem.setRotationType(static_cast<RotationType>(rotationType));
     auto d = rotatedItem.getDimension();
@@ -29,27 +29,26 @@ float Bin::scoreRotation(const Item& item, int rotationType) const {
     if (getWidth() < d[0] || getHeight() < d[1] || getDepth() < d[2]) {
         return 0;
     }
-
-    float widthScore = std::pow(d[0] / getWidth(), 2);
-    float heightScore = std::pow(d[1] / getHeight(), 2);
-    float depthScore = std::pow(d[2] / getDepth(), 2);
+    float widthScore = std::pow(static_cast<float>(d[0]) / getWidth(), 2);
+    float heightScore = std::pow(static_cast<float>(d[1]) / getHeight(), 2);
+    float depthScore = std::pow(static_cast<float>(d[2]) / getDepth(), 2);
 
     float score = widthScore + heightScore + depthScore;
     return score;
 }
 
-std::vector<int> Bin::getBestRotationOrder(const Item& item) const {
+std::vector<long> Bin::getBestRotationOrder(const Item& item) const {
     // Use map to maintain same structure as Python dictionary
-    std::map<int, float> rotationScores;
+    std::map<long, float> rotationScores;
     // Score all rotation types
     for (auto rotation : item.getAllowedRotations()) {
-        // Explicitly cast RotationType to int for map key
-        int rotationInt = static_cast<int>(rotation);
+        // Explicitly cast RotationType to long for map key
+        long rotationInt = static_cast<long>(rotation);
         rotationScores[rotationInt] = scoreRotation(item, rotationInt);
     }
     
     // Convert map to vector for sorting
-    std::vector<int> bestRotations;
+    std::vector<long> bestRotations;
     bestRotations.reserve(rotationScores.size());
     
     // Extract keys and sort them based on their corresponding values
@@ -58,19 +57,19 @@ std::vector<int> Bin::getBestRotationOrder(const Item& item) const {
     }
     
     std::sort(bestRotations.begin(), bestRotations.end(),
-              [&rotationScores](int a, int b) {
+              [&rotationScores](long a, long b) {
                   return rotationScores[a] > rotationScores[b];
               });
     
     return bestRotations;
 }
 
-bool Bin::putItem(Item& item, const std::tuple<float, float, float>& p) {
+bool Bin::putItem(Item& item, const std::tuple<long, long, long>& p) {
     bool fit = false;
     auto rotations = getBestRotationOrder(item);
 
     item.setPosition({std::get<0>(p), std::get<1>(p), std::get<2>(p)});
-    for (int rotation : rotations) {
+    for (long rotation : rotations) {
         item.setRotationType(static_cast<RotationType>(rotation));
         auto d = item.getDimension();
