+++ src/box.cpp
@@ -1,10 +1,10 @@
 #include "box.h"
 #include "utils.h"
 
-Box::Box(const std::string& name, float width, float height, float depth)
+Box::Box(const std::string& name, long width, long height, long depth)
     : name(name), width(factored_integer(width)), height(factored_integer(height)), depth(factored_integer(depth)) {}
 
-float Box::getVolume() const {
+long Box::getVolume() const {
     return width * height * depth;
 }
 
@@ -12,14 +12,14 @@ std::string Box::getName() const {
     return name;
 }
 
-float Box::getWidth() const {
+long Box::getWidth() const {
     return width;
 }
 
-float Box::getHeight() const {
+long Box::getHeight() const {
     return height;
 }
 
-float Box::getDepth() const {
+long Box::getDepth() const {
     return depth;
 }
