+++ src/utils.cpp
@@ -1,8 +1,8 @@
 #include "utils.h"
 #include <cmath>
 
-const int FACTOR = 0;
+const long FACTOR = 0;
 
-int factored_integer(double value) {
-    return static_cast<int>(round(value * pow(10, FACTOR)));
+long factored_integer(double value) {
+    return static_cast<long>(round(value * pow(10, FACTOR)));
 }
